// src/features/auth/infrastructure/api/model/sectionEntity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Tickit Server API
 * Tickit 클라우드 서버 API 명세서
 * OpenAPI spec version: 1.0
 */
import type { ReminderEntity } from './reminderEntity';

export interface SectionEntity {
  /** 섹션 고유 ID */
  id: string;
  /** 섹션 제목 */
  title: string;
  /** 고정 섹션 여부 */
  isFixed: boolean;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  updatedAt: string;
  reminders?: ReminderEntity[];
}